requirements.txt
# Web框架
Flask==2.3.2
Werkzeug==2.3.6
flask-cors==4.0.0

# 异步HTTP客户端
aiohttp==3.8.5
aiofiles==23.2.1

# HTTP客户端
requests==2.31.0

# LLM接口
anthropic==0.18.1

# 图像处理
Pillow==10.1.0
numpy==1.24.3

# 环境变量管理
python-dotenv==1.0.0

# 模板引擎
jinja2==3.1.2